docs: Make README links work if viewed in code repository
An unexpected feature of the README-to-GitHub Pages conversion is that
any link in the docs/README.md that refers to a **generated** file, will
of course appear as 404 if it was clicked from within the code
repository.

The root of the GitHub pages is generated from README.md, which is of
course easy to find (and even displayed automatically by GitHub while
navigating the code tree into the docs directory).

Generated files are not stored in the repository, so those links only
work when the README is viewed as part of the GitHub Pages.
The old style works from within the GitHub pages where the generated
files actually exist.

Specifying an absolute URL for each one feels less future-proof than a
relative path, but it needs to be done so that there are not broken
links for anyone who finds the README via the code repository tree
instead of the GitHub Pages.

The links to files that are static (exist already under docs) were
left as relative links, since they work in both situations.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>